+++ engine/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.34.0
 ### Features
 * Published the `DefaultHttpLogsCollectorPortNum` const which can be used to launch the engine's server